requirements.txt
Flask==3.0.0
opencv-python==4.9.0.80
pillow==10.2.0